moved class Api to folder API
