+++ source/php/Employee/EmployeeApiManager.php
@@ -2,7 +2,7 @@
 
 namespace VolunteerManager\Employee;
 
-use VolunteerManager\Api;
+use VolunteerManager\API\Api;
 use VolunteerManager\API\WPResponseFactory;
 use WP_Error;
 use WP_REST_Request;
